Update library versions in build scripts
This commit updates several dependencies in the project, including the Kotlin version to "2.1.0", the Ktor version to "3.0.1", and various others to their respective latest stable releases. Additionally, the library version in the Gradle build script is incremented from "1.2.3" to "1.2.4". These changes aim to ensure compatibility with the latest features and improvements.